refactor: Refactored tests
Moved tests for each app to seperate test files under tests folder
