+++ bug_tracker/bugs/tests/test_models.py
@@ -1,9 +1,7 @@
 from django.contrib.auth import get_user_model
 from django.test import TestCase
-from django.urls import reverse
-from users.models import CustomUser
 
-from .models import Bug
+from ..models import Bug
 
 
 class BugModelTest(TestCase):
@@ -124,97 +122,3 @@ class BugModelTest(TestCase):
         bug = Bug.objects.get(pk=1)
 
         self.assertEqual(str(bug), "A Title")
-
-
-class BugListViewTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        User = get_user_model()
-
-        User.objects.create_user(
-            email="normal@test.com", first_name="John", last_name="Doe", password="test"
-        )
-
-    def test_view_logged_in(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get("")
-
-        self.assertEqual(response.status_code, 200)
-
-    def test_view_not_logged_in(self):
-        response = self.client.get("")
-
-        self.assertRedirects(response, "/login/?next=/")
-
-    def test_view_url_accessible_by_name(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_list"))
-
-        self.assertEqual(response.status_code, 200)
-
-    def test_view_uses_correct_template(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_list"))
-
-        self.assertTemplateUsed(response, "bugs/list.html")
-
-
-class BugCreateViewTest(TestCase):
-    @classmethod
-    def setUpTestData(cls):
-        User = get_user_model()
-
-        User.objects.create_user(
-            email="normal@test.com", first_name="John", last_name="Doe", password="test"
-        )
-
-        manager: CustomUser = User.objects.create_user(
-            email="manager@test.com",
-            first_name="Jane",
-            last_name="Doe",
-            password="test",
-        )
-
-        manager.user_type = CustomUser.MANAGER
-        manager.save()
-
-    def test_view_logged_in(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get("/create/")
-
-        self.assertEqual(response.status_code, 200)
-
-    def test_view_not_logged_in(self):
-        response = self.client.get("/create/")
-
-        self.assertRedirects(response, "/login/?next=/create/")
-
-    def test_view_url_accessible_by_name(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_create"))
-
-        self.assertEqual(response.status_code, 200)
-
-    def test_view_uses_correct_template(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_create"))
-
-        self.assertTemplateUsed(response, "bugs/create.html")
-
-    def test_view_form_manager(self):
-        self.client.login(username="manager@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_create"))
-
-        self.assertContains(response, "Assignee")
-
-    def test_view_from_contains_assignee_when_manager(self):
-        self.client.login(username="manager@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_create"))
-
-        self.assertContains(response, "Assignee")
-
-    def test_view_from_does_not_contains_assignee_when_developer(self):
-        self.client.login(username="normal@test.com", password="test")
-        response = self.client.get(reverse("bugs:bug_create"))
-
-        self.assertNotContains(response, "Assignee")
